improve: add react router and landing page
- update lint rule
