+++ tslint.json
@@ -25,8 +25,7 @@
       true,
       "parameter",
       "arrow-parameter",
-      "property-declaration",
-      "variable-declaration"
+      "property-declaration"
     ],
     "typedef-whitespace": [
       true,
